UserPersistenceManager overhaul
